+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 1.1.1
 
 - No longer requiring chrisboulton/php-resque-scheduler, rather suggesting it.
+- Added the resque:listen Artisan command.
 
 ### 1.1.0